move getRuntimeVersion to config-plugins
