--- packages/config/src/__tests__/Project-test.ts
@@ -1,104 +1,0 @@
-import {
-  getBuildNumber,
-  getNativeVersion,
-  getRuntimeVersion,
-  getVersion,
-  getVersionCode,
-} from '../Project';
-
-console.warn = jest.fn();
-
-describe(getNativeVersion, () => {
-  const version = '2.0.0';
-  const versionCode = 42;
-  const buildNumber = '13';
-  it('works for android', () => {
-    expect(getNativeVersion({ version, android: { versionCode } }, 'android')).toBe(
-      `${version}(${versionCode})`
-    );
-  });
-  it('works for ios', () => {
-    expect(getNativeVersion({ version, ios: { buildNumber } }, 'ios')).toBe(
-      `${version}(${buildNumber})`
-    );
-  });
-  it('throws an error if platform is not recognized', () => {
-    const fakePlatform = 'doesnotexit';
-    expect(() => {
-      getNativeVersion({ version }, fakePlatform as any);
-    }).toThrow(`"${fakePlatform}" is not a supported platform. Choose either "ios" or "android".`);
-  });
-  it('uses the default version if the version is missing', () => {
-    expect(getNativeVersion({}, 'ios')).toBe('1.0.0(1)');
-  });
-  it('uses the default buildNumber if the platform is ios and the buildNumber is missing', () => {
-    expect(getNativeVersion({ version }, 'ios')).toBe(`${version}(1)`);
-  });
-  it('uses the default versionCode if the platform is android and the versionCode is missing', () => {
-    expect(getNativeVersion({ version }, 'android')).toBe(`${version}(1)`);
-  });
-});
-
-describe(getVersion, () => {
-  it(`uses version if it's given in config`, () => {
-    expect(getVersion({ version: '1.2.3' })).toBe('1.2.3');
-  });
-
-  it(`uses 1.0.0 if no version is given`, () => {
-    expect(getVersion({})).toBe('1.0.0');
-  });
-});
-describe(getVersionCode, () => {
-  it(`returns 1 if no version code is provided`, () => {
-    expect(getVersionCode({})).toBe(1);
-  });
-
-  it(`returns the version code if provided`, () => {
-    expect(getVersionCode({ android: { versionCode: 5 } })).toBe(5);
-  });
-});
-describe(getBuildNumber, () => {
-  it(`uses ios.buildNumber if it's given in config`, () => {
-    expect(getBuildNumber({ ios: { buildNumber: '12' } })).toEqual('12');
-  });
-
-  it(`falls back to '1' if not provided`, () => {
-    expect(getBuildNumber({ ios: { buildNumber: null } })).toEqual('1');
-    expect(getBuildNumber({})).toEqual('1');
-  });
-});
-describe(getRuntimeVersion, () => {
-  it('works if the top level runtimeVersion is a string', () => {
-    const runtimeVersion = '42';
-    expect(getRuntimeVersion({ runtimeVersion }, 'ios')).toBe(runtimeVersion);
-  });
-  it('works if the platform specific runtimeVersion is a string', () => {
-    const runtimeVersion = '42';
-    expect(getRuntimeVersion({ ios: { runtimeVersion } }, 'ios')).toBe(runtimeVersion);
-  });
-  it('works if the runtimeVersion is a policy', () => {
-    const version = '1';
-    const buildNumber = '2';
-    expect(
-      getRuntimeVersion(
-        { version, runtimeVersion: { policy: 'nativeVersion' }, ios: { buildNumber } },
-        'ios'
-      )
-    ).toBe(`${version}(${buildNumber})`);
-  });
-  it('returns default value if no runtime version is supplied', () => {
-    expect(getRuntimeVersion({}, 'ios')).toBe('1.0.0(1)');
-  });
-  it('throws if runtime version is not parseable', () => {
-    expect(() => {
-      getRuntimeVersion({ runtimeVersion: 1 } as any, 'ios');
-    }).toThrow(
-      `"1" is not a valid runtime version. getRuntimeVersion only supports a string or the "nativeVersion" policy.`
-    );
-    expect(() => {
-      getRuntimeVersion({ runtimeVersion: { policy: 'unsupportedPlugin' } } as any, 'ios');
-    }).toThrow(
-      `"{"policy":"unsupportedPlugin"}" is not a valid runtime version. getRuntimeVersion only supports a string or the "nativeVersion" policy.`
-    );
-  });
-});
